+++ lib/code_builder.dart
@@ -10,7 +10,13 @@ export 'src/builders/expression.dart'
 export 'src/builders/field.dart'
     show varConst, varField, varFinal, FieldBuilder;
 export 'src/builders/file.dart'
-    show ExportBuilder, ImportBuilder, LibraryBuilder, PartBuilder;
+    show
+        ExportBuilder,
+        ImportBuilder,
+        LibraryBuilder,
+        PartBuilder,
+        PartOfBuilder,
+        UriDirectiveBuilder;
 export 'src/builders/method.dart'
     show
         constructor,
@@ -25,7 +31,8 @@ export 'src/builders/method.dart'
         MethodBuilder,
         MethodModifier,
         ValidMethodMember;
-export 'src/builders/parameter.dart' show parameter, ParameterBuilder;
+export 'src/builders/parameter.dart'
+    show parameter, FunctionParameterBuilder, ParameterBuilder;
 export 'src/pretty_printer.dart' show prettyToSource;
 export 'src/builders/reference.dart'
     show explicitThis, reference, ReferenceBuilder;
@@ -47,4 +54,5 @@ export 'src/builders/statement.dart'
         SwitchDefaultCaseBuilder,
         SwitchStatementBuilder,
         WhileStatementBuilder;
-export 'src/builders/type.dart' show NewInstanceBuilder, TypeBuilder;
+export 'src/builders/type.dart'
+    show NewInstanceBuilder, TypeBuilder, TypeDefBuilder;