+++ test/builders/expression_test.dart
@@ -29,8 +29,30 @@ void main() {
       expect(literal(5.5), equalsSource(r'5.5'));
     });
 
-    test('should emit a string', () {
-      expect(literal('Hello'), equalsSource(r"'Hello'"));
+    group('should emit a string', () {
+      test('simple', () {
+        expect(literal('Hello'), equalsSource(r"'Hello'"));
+      });
+
+      test("with a '", () {
+        expect(literal("Hello'"), equalsSource(r"'Hello\''"));
+      });
+
+      test(r"with a \'", () {
+        expect(literal(r"Hello\'"), equalsSource(r"'Hello\\\''"));
+      });
+
+      test(r"with a \\'", () {
+        expect(literal(r"Hello\\'"), equalsSource(r"'Hello\\\\\''"));
+      });
+
+      test(r"with a newline", () {
+        expect(literal("Hello\nworld"), equalsSource("'''Hello\nworld'''"));
+      });
+
+      test(r"with a newline and ending with a '", () {
+        expect(literal("Hello\nworld'"), equalsSource("'''Hello\nworld\\''''"));
+      });
     });
 
     test('should emit a list', () {
@@ -401,6 +423,26 @@ void main() {
       literal(1.0).castAs(lib$core.num),
       equalsSource(r'''
         1.0 as num
+      '''),
+    );
+  });
+
+  test('should throw an exception', () {
+    expect(
+        new TypeBuilder('StateError')
+            .newInstance([literal('Hey! No!')])
+            .asThrow()
+            .asStatement(),
+        equalsSource('''
+    throw new StateError('Hey! No!');
+    '''));
+  });
+
+  test('should create a ternary condition ? ifTrue : ifFalse', () {
+    expect(
+      reference('someValue').ternary(literal(true), literal(false)),
+      equalsSource(r'''
+        someValue ? true : false
       '''),
     );
   });